Added role based functionality for UserService and TeamService.

+++ TeamManagement/src/main/java/pl/rynski/teammanagement/repository/TeamRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.rynski.teammanagement.model.Team;
 
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 
 }
